third commit

+++ resources/views/nilai/index.blade.php
@@ -3,7 +3,9 @@
     <center>
         <br>
             <h2>LIST DATA NILAI</h2>
-            <a href="/nilai/create" class="button-primary">TAMBAH DATA</a>
+            @if (session('user')->role == 'guru')
+                <a href="/nilai/create" class="button-primary">TAMBAH DATA</a>
+            @endif
             @if (session('success'))
                 <p class="text-success">{{ session('success') }}</p>
             @endif
@@ -19,9 +21,9 @@
                     <th>UTS</th>
                     <th>UAS</th>
                     <th>NA</th>
-                    {{-- @if (session('user')->role == "guru") --}}
+                    @if (session('user')->role == "guru")
                         <th>ACTION</th>
-                    {{-- @endif --}}
+                    @endif
                 </tr>
                 @foreach ($nilai as $n)
                     <tr>
@@ -32,12 +34,12 @@
                         <td>{{ $n ->uts }}</td>
                         <td>{{ $n ->uas }}</td>
                         <td>{{ $n ->na }}</td>
-                        {{-- @if (session('user')->role == "guru") --}}
+                        @if (session('user')->role == "guru")
                             <td>
                                 <a href="/nilai/edit/{{ $n -> id }}" class="button-warning">Edit</a>
                                 <a href="/nilai/destroy/{{ $n -> id }}" class="button-danger" onclick="return confrim('Yakin ingin dihapus?')">Hapus</a>
                             </td>
-                        {{-- @endif --}}
+                        @endif
                     </tr>
                 @endforeach
             </table>
